+++ libbpf-tools/opensnoop.c
@@ -8,6 +8,7 @@
 #define _GNU_SOURCE
 #endif
 #include <argp.h>
+#include <inttypes.h>
 #include <fcntl.h>
 #include <signal.h>
 #include <stdio.h>
@@ -307,7 +308,7 @@ int handle_event(void *ctx, void *data, size_t data_sz)
 
 void handle_lost_events(void *ctx, int cpu, __u64 lost_cnt)
 {
-	fprintf(stderr, "Lost %llu events on CPU #%d!\n", lost_cnt, cpu);
+	fprintf(stderr, "Lost %" PRIu64" events on CPU #%d!\n", (uint64_t)lost_cnt, cpu);
 }
 
 int main(int argc, char **argv)
